Remove `-s` in favor of `active_store` setting
When you are making management commands, it's
annoying to have to type `-s store` after every
command when you are often trying to run many
commands on the same store. Now you can set an
active store with the `settings` command group and
that store will be used for all requests.