Use `click.edit` for put commands
Previously, I was rolling my own solution for the
data editing in put commands that I was not really
happy with. My main goal was to allow VIM, which
is much better than nano for this type of thing.
When reading I found out that click had a built
in edit command and it seemed like a no-brainer to
use that.